Split Snapshot overlay by metric aggregation type (#35)
* Split Snapshot overlay by metric aggregation type

Event based metrics like `resignationCount` does aggregate across time,
i.e. it is cumulative. This is different from e.g. `employeeCount`,
which is a Subject-based metric and is non-cumulative over time (similar
in concept to e.g. inventory level).

This means that the Overlays that capture the snapshot values of these
metrics must match their aggregation semantics. Because the metrics are
different in this regard, the overlays have also then to be distinct.

* update readme

* Doc review/split snapshot overlay by metric aggregation type (#36)

* Update README.md

* Update gender-headcount-joblevel.json

Fix Headcount spelling

* Update README.md

* Update gender-headcount-joblevel.json

revert

---------

Co-authored-by: emily-watt-visier <141189938+emily-watt-visier@users.noreply.github.com>